Update following posts component user experience
Refactor follow service so following usernames array is the only field
holding that array data and is manipulated in place by follow component.
Create new behavior subject that stores an unfollowed username, so
following posts component can perform some actions when a new unfollow
is detected. PostRes model was updated to also animateDelete so classes
could be dynamically applied when a post is to be deleted.